requirements.txt
opencv-python>=4.8.0,<5
mediapipe>=0.10.9,<0.11
numpy>=1.24.0,<3
Pillow>=10.0.0
scikit-learn>=1.3.0
onnxruntime>=1.16.0
